Add swipe, drag, scroll functions for slideshow
Users can now swipe through each slide in the slideshow by swiping with
their finger on mobile devices, clicking and dragging left or right, or
scrolling left or right on a touchpad. This is in addition to the
previous functionality which played the slides automatically, and allows
the user to click the manual button controls on the bottom.