Add pages folder, organize .js files, import font
I've added the html files for each page and added their respective links
to the navigation menu so users can access all pages.

I changed how the .js files were organized, and linked up the pages to
the applicable scripts. Global.js is for script that will be implemented
on all pages. index.js is for the index page only.

The 'CelloSansLight' font stopped responding, and the website I imported
from can't be found, so I downloaded the font, created a font folder,
and moved them there, then imported it into styles.css. Currently, only
CelloSans-ExtraLight.ttf is in use, but I added the others in case I
need to use them in the future. Once this website is complete, I plan to
remove any unnecessary .ttf files.