Rename things for clarity, add container function
I created a function to handle artwork that is meant to be viewed as
a group of images, like mini-comics. This is so users will be able to
tell a clear difference between singular artworks and comics by adding
visuals to indicate which artwork is a comic, as well as how many pages
that comic is. I've renamed certain variables and functions to have more
unity and clarity so it's easier to read.