Create filter and populate filter options
The only thing the filter does right now is allow users to change the
order from newest-oldest and vice versa. The createFilter function
populates the drop-down menus with options taken from the art objects'
year and series properties. This is so the filter is automatically
updated whenever new art is added.