Add scroll/swipe/grabbing for reading comics
I added functions to instead of scrolling through comic pages, for users
to swipe, scroll, or grab and drag. This will smoothly translate the
comic to the next page, rather than users having to scroll through the
pages, which meant reading the comics in an unintended way. I've added
stronger indication that comics can be swiped through by adding an
animation that can be smoothly interrupted.